+++ code_base/micro-services/desk-service/src/main/java/github/opendesk/deskservice/config/DeskConfig.java
@@ -1,5 +1,6 @@
 package github.opendesk.deskservice.config;
 
+import org.springframework.amqp.core.*;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
@@ -20,4 +21,45 @@ public class DeskConfig {
     RestTemplate restTemplate() {
         return new RestTemplate();
     }
+
+    /**
+     * Rabbitmq
+     */
+
+    @Bean
+    public FanoutExchange fanout() {
+        return new FanoutExchange("fanout");
+    }
+
+    @Bean
+    public Queue autoDeleteQueue1() {
+        return new AnonymousQueue();
+    }
+
+    @Bean
+    public Queue autoDeleteQueue2() {
+        return new AnonymousQueue();
+    }
+
+    @Bean
+    public Binding binding1(FanoutExchange fanout,
+                            Queue autoDeleteQueue1) {
+        return BindingBuilder.bind(autoDeleteQueue1).to(fanout);
+    }
+
+    @Bean
+    public Binding binding2(FanoutExchange fanout,
+                            Queue autoDeleteQueue2) {
+        return BindingBuilder.bind(autoDeleteQueue2).to(fanout);
+    }
+
+    @Bean
+    public RabbitMQReceiver receiver() {
+        return new RabbitMQReceiver();
+    }
+
+
+
+
+
 }
