+++ code_base/micro-services/desk-service/src/main/java/github/opendesk/deskservice/config/RabbitMQReceiver.java
@@ -2,25 +2,20 @@ package github.opendesk.deskservice.config;
 import org.springframework.amqp.core.*;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
+import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.util.StopWatch;
 
 @Configuration
 public class RabbitMQReceiver {
-    @Bean
-    public FanoutExchange fanout() {
-        return new FanoutExchange("fanout");
+
+    @RabbitListener(queues = "#{autoDeleteQueue1.name}")
+    public void receive1(String in) throws InterruptedException {
+        receive(in, 1);
     }
 
-    @Bean
-    public Queue autoDeleteQueue() {
-        return new AnonymousQueue();
-    }
-
-    @Bean
-    public Binding binding(FanoutExchange fanout,
-                            Queue autoDeleteQueue) {
-        return BindingBuilder.bind(autoDeleteQueue).to(fanout);
+    @RabbitListener(queues = "#{autoDeleteQueue2.name}")
+    public void receive2(String in) throws InterruptedException {
+        receive(in, 2);
     }
 
     public void receive(String in, int receiver) throws InterruptedException {
@@ -28,11 +23,7 @@ public class RabbitMQReceiver {
         watch.start();
         System.out.println("instance " + receiver + " [x] Received '" + in + "'");
         watch.stop();
-    }
-
-    @Bean
-    public RabbitMQReceiver receiver() {
-        return new RabbitMQReceiver();
+        System.out.println("instance " + receiver + " [x] Done in " + watch.getTotalTimeSeconds() + "s");
     }
 
 }
