+++ code_base/micro-services/organisation-service/src/main/java/github/opendesk/organisationservice/config/RabbitMQSender.java
@@ -3,6 +3,7 @@ package github.opendesk.organisationservice.config;
 import org.springframework.amqp.core.FanoutExchange;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.context.annotation.Bean;
 
@@ -14,14 +15,8 @@ public class RabbitMQSender {
     @Autowired
     private FanoutExchange fanout;
 
-    @Bean
-    public RabbitMQSender sender() {
-        return new RabbitMQSender();
-    }
-
-    @Scheduled(fixedDelay = 1000, initialDelay = 500)
     public void send(String message) {
-        this.template.convertAndSend(fanout.getName(), message);
+        template.convertAndSend(fanout.getName(), "", message);
         System.out.println(" Sent '" + message + "'");
     }
 }
